Factor out session manager to shared location.
