+++ tests/integration/test_utils/data_source_config/big_query.py
@@ -9,6 +9,7 @@ from great_expectations.compatibility.pydantic import BaseSettings
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.data_context import AbstractDataContext
 from great_expectations.datasource.fluent.sql_datasource import TableAsset
+from tests.integration.sql_session_manager import SessionSQLEngineManager
 from tests.integration.test_utils.data_source_config.base import (
     BatchTestSetup,
     DataSourceTestConfig,
@@ -17,8 +18,6 @@ from tests.integration.test_utils.data_source_config.sql import SQLBatchTestSetu
 
 if TYPE_CHECKING:
     import pandas as pd
-
-    from tests.integration.conftest import TestSessionSQLEngineManager
 
 
 class BigQueryDatasourceTestConfig(DataSourceTestConfig):
@@ -39,7 +38,7 @@ class BigQueryDatasourceTestConfig(DataSourceTestConfig):
         data: pd.DataFrame,
         extra_data: Mapping[str, pd.DataFrame],
         context: AbstractDataContext,
-        engine_manager: Optional[TestSessionSQLEngineManager] = None,
+        engine_manager: Optional[SessionSQLEngineManager] = None,
     ) -> BatchTestSetup:
         return BigQueryBatchTestSetup(
             data=data,
