+++ tests/integration/test_utils/data_source_config/databricks.py
@@ -10,6 +10,7 @@ from great_expectations.compatibility.sqlalchemy import sqltypes
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.data_context import AbstractDataContext
 from great_expectations.datasource.fluent.sql_datasource import TableAsset
+from tests.integration.sql_session_manager import SessionSQLEngineManager
 from tests.integration.test_utils.data_source_config.base import (
     BatchTestSetup,
     DataSourceTestConfig,
@@ -21,8 +22,6 @@ from tests.integration.test_utils.data_source_config.sql import (
 
 if TYPE_CHECKING:
     import pandas as pd
-
-    from tests.integration.conftest import TestSessionSQLEngineManager
 
 
 class DatabricksDatasourceTestConfig(DataSourceTestConfig):
@@ -43,7 +42,7 @@ class DatabricksDatasourceTestConfig(DataSourceTestConfig):
         data: pd.DataFrame,
         extra_data: Mapping[str, pd.DataFrame],
         context: AbstractDataContext,
-        engine_manager: Optional[TestSessionSQLEngineManager] = None,
+        engine_manager: Optional[SessionSQLEngineManager] = None,
     ) -> BatchTestSetup:
         return DatabricksBatchTestSetup(
             data=data,
