+++ tests/integration/test_utils/data_source_config/mssql.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Mapping, Optional
+from typing import Mapping, Optional
 
 import pandas as pd
 import pytest
@@ -6,15 +6,13 @@ import pytest
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.data_context import AbstractDataContext
 from great_expectations.datasource.fluent.sql_datasource import TableAsset
+from tests.integration.sql_session_manager import SessionSQLEngineManager
 from tests.integration.test_utils.data_source_config.base import (
     BatchTestSetup,
     DataSourceTestConfig,
 )
 from tests.integration.test_utils.data_source_config.sql import SQLBatchTestSetup
 from tests.test_utils import get_default_mssql_url
-
-if TYPE_CHECKING:
-    from tests.integration.conftest import TestSessionSQLEngineManager
 
 
 class MSSQLDatasourceTestConfig(DataSourceTestConfig):
@@ -35,7 +33,7 @@ class MSSQLDatasourceTestConfig(DataSourceTestConfig):
         data: pd.DataFrame,
         extra_data: Mapping[str, pd.DataFrame],
         context: AbstractDataContext,
-        engine_manager: Optional[TestSessionSQLEngineManager] = None,
+        engine_manager: Optional[SessionSQLEngineManager] = None,
     ) -> BatchTestSetup:
         return MSSQLBatchTestSetup(
             data=data,
