+++ tests/integration/test_utils/data_source_config/mysql.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Mapping, Optional
+from typing import Mapping, Optional
 
 import pandas as pd
 import pytest
@@ -7,6 +7,7 @@ from great_expectations.compatibility.sqlalchemy import sqltypes
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.data_context import AbstractDataContext
 from great_expectations.datasource.fluent.sql_datasource import TableAsset
+from tests.integration.sql_session_manager import SessionSQLEngineManager
 from tests.integration.test_utils.data_source_config.base import (
     BatchTestSetup,
     DataSourceTestConfig,
@@ -15,9 +16,6 @@ from tests.integration.test_utils.data_source_config.sql import (
     InferrableTypesLookup,
     SQLBatchTestSetup,
 )
-
-if TYPE_CHECKING:
-    from tests.integration.conftest import TestSessionSQLEngineManager
 
 
 class MySQLDatasourceTestConfig(DataSourceTestConfig):
@@ -38,7 +36,7 @@ class MySQLDatasourceTestConfig(DataSourceTestConfig):
         data: pd.DataFrame,
         extra_data: Mapping[str, pd.DataFrame],
         context: AbstractDataContext,
-        engine_manager: Optional[TestSessionSQLEngineManager] = None,
+        engine_manager: Optional[SessionSQLEngineManager] = None,
     ) -> BatchTestSetup:
         return MySQLBatchTestSetup(
             data=data,
