+++ tests/integration/test_utils/data_source_config/pandas_filesystem_csv.py
@@ -1,6 +1,6 @@
 import pathlib
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Any, Mapping, Optional
+from typing import Any, Mapping, Optional
 
 import pandas as pd
 import pytest
@@ -9,13 +9,11 @@ from great_expectations.compatibility.typing_extensions import override
 from great_expectations.data_context import AbstractDataContext
 from great_expectations.datasource.fluent.data_asset.path.pandas.generated_assets import CSVAsset
 from great_expectations.datasource.fluent.interfaces import Batch
+from tests.integration.sql_session_manager import SessionSQLEngineManager
 from tests.integration.test_utils.data_source_config.base import (
     BatchTestSetup,
     DataSourceTestConfig,
 )
-
-if TYPE_CHECKING:
-    from tests.integration.conftest import TestSessionSQLEngineManager
 
 
 @dataclass(frozen=True)
@@ -42,7 +40,7 @@ class PandasFilesystemCsvDatasourceTestConfig(DataSourceTestConfig):
         data: pd.DataFrame,
         extra_data: Mapping[str, pd.DataFrame],
         context: AbstractDataContext,
-        engine_manager: Optional[TestSessionSQLEngineManager] = None,
+        engine_manager: Optional[SessionSQLEngineManager] = None,
     ) -> BatchTestSetup:
         assert not extra_data, "extra_data is not supported for this data source."
         tmp_path = request.getfixturevalue("tmp_path")
