+++ tests/integration/test_utils/data_source_config/sql.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 
-import logging
 from abc import ABC, abstractmethod
-from dataclasses import asdict, dataclass
+from dataclasses import dataclass
 from datetime import date, datetime
 from functools import cached_property
 from typing import Any, Callable, Generic, Mapping, Optional, Sequence, Union
@@ -10,7 +9,6 @@ from typing import Any, Callable, Generic, Mapping, Optional, Sequence, Union
 import numpy as np
 import pandas as pd
 import sqlalchemy as sa
-from sqlalchemy.pool import QueuePool
 from typing_extensions import override
 
 from great_expectations.compatibility.sqlalchemy import (
@@ -26,92 +24,11 @@ from great_expectations.compatibility.sqlalchemy import (
 from great_expectations.data_context import AbstractDataContext
 from great_expectations.datasource.fluent.interfaces import Batch
 from great_expectations.datasource.fluent.sql_datasource import TableAsset
+from tests.integration.sql_session_manager import (
+    ConnectionDetails,
+    SessionSQLEngineManager,
+)
 from tests.integration.test_utils.data_source_config.base import BatchTestSetup, _ConfigT
-
-logger = logging.getLogger(__name__)
-
-
-@dataclass(frozen=True)
-class ConnectionDetails:
-    connection_string: str
-    dialect: str
-
-
-@dataclass(frozen=True)
-class PoolConfig:
-    pool_size: int
-    max_overflow: int
-    pool_recycle: int
-    pool_timeout: int
-    pool_pre_ping: bool
-
-
-class TestSessionSQLEngineManager:
-    POOL_CONFIG = PoolConfig(
-        pool_size=2,
-        max_overflow=3,
-        pool_recycle=5400,  # 1.5 hours
-        pool_timeout=30,  # 30 seconds
-        pool_pre_ping=True,
-    )
-
-    def __init__(self):
-        self._engine_cache: dict[ConnectionDetails, sa.engine.Engine] = {}
-
-    def get_engine(
-        self,
-        connection_details: ConnectionDetails,
-    ) -> sa.engine.Engine:
-        cache_key = connection_details
-        if cache_key not in self._engine_cache:
-            logger.info(f"Cache miss for engine: {cache_key}. Creating new engine.")
-            engine_kwargs = asdict(self.POOL_CONFIG)
-            logger.info(
-                f"Creating engine for {connection_details.dialect} with settings: {engine_kwargs}"
-            )
-            self._engine_cache[cache_key] = sa.create_engine(
-                connection_details.connection_string, **engine_kwargs
-            )
-        else:
-            logger.info(f"Cache hit for engine: {cache_key}")
-        return self._engine_cache[cache_key]
-
-    def dispose_all_engines(self):
-        logger.info("Disposing all cached SQLAlchemy engines.")
-        for key, engine in self._engine_cache.items():
-            logger.info(f"Disposing engine: {key}")
-            try:
-                engine.dispose()
-            except Exception:
-                logger.exception(f"Error disposing engine '{key}'")
-        self._engine_cache.clear()
-
-    def get_all_pool_statistics(
-        self,
-    ) -> dict[ConnectionDetails, dict[str, Any]]:
-        stats: dict[ConnectionDetails, dict[str, Any]] = {}
-        for key, engine in self._engine_cache.items():
-            try:
-                pool = engine.pool
-                if isinstance(pool, QueuePool):
-                    stats[key] = {
-                        "size": pool.size(),
-                        "checked_in": pool.checkedin(),
-                        "overflow": pool.overflow(),
-                        "checked_out": pool.checkedout(),
-                    }
-                else:
-                    logger.warning(
-                        f"Pool for engine {key} is not a QueuePool. It is a {type(pool)}."
-                    )
-                    stats[key] = {
-                        "type": f"{type(pool)}",
-                        "status": f"{pool.status()}",
-                    }
-            except Exception as e:
-                logger.exception(f"Error getting pool status for engine '{key}'")
-                stats[key] = {"error": str(e)}
-        return stats
 
 
 @dataclass(frozen=True)
@@ -163,7 +80,7 @@ class SQLBatchTestSetup(BatchTestSetup[_ConfigT, TableAsset], ABC, Generic[_Conf
         extra_data: Mapping[str, pd.DataFrame],
         context: AbstractDataContext,
         table_name: Optional[str] = None,  # Overrides random table name generation
-        engine_manager: Optional[TestSessionSQLEngineManager] = None,
+        engine_manager: Optional[SessionSQLEngineManager] = None,
     ) -> None:
         self.engine_manager = engine_manager
         self.extra_data = extra_data
@@ -219,7 +136,7 @@ class SQLBatchTestSetup(BatchTestSetup[_ConfigT, TableAsset], ABC, Generic[_Conf
         if self.engine_manager:
             connection_details = ConnectionDetails(
                 connection_string=self.connection_string,
-                dialect=self.config.label,  # The config label should be the sqlalchemy dialect
+                dialect=sa.dialects.get_dialect(self.connection_string).name,
             )
             engine = self.engine_manager.get_engine(connection_details)
             return engine, lambda: None
